requirements.txt
PyYAML
python-telegram-bot[ext]
ptbcontrib[ptb_jobstores_sqlalchemy] @git+https://github.com/python-telegram-bot/ptbcontrib.git@main
SQLAlchemy~=2.0.27
